Home Page Images
I styled the main images to be rows and colums so they sit next to each other. I think I will need to resize images to make them even so keep that in mind as you add images to your pages. I also made the images links to each of your pages when you click on them; this does not affect the stling at all.