Sytling Header, Navagation, and Home page
Imgas on the home page are sized indiviualy under main in the css. This means all your images will need to be put under their own div and styled in the css. I added a gif for the header and nav buttons. I am playing with the sizing on the nav buttons to make it more visable.