NinjaBrain Sytling
I corrected the HTML to be h1, h2, h3, and p since they were styled. I added the images and put the remaining content on the page. I added links to a vidoe thumbnail rather than a simple link. I also made new folders to organize the Images. I have begun some work on the Manual Page but not much. It will have the same formating and styling as Ninja brain with differing images and no downlaod.